+++ PlantsVs/src/atlas.cpp
@@ -3,10 +3,10 @@
 #include "utils/graphics_util.h"
 
 namespace PlantsVs {
-void Atlas::LoadFromFile(std::string path, int num, std::string suffix) {
+void Atlas::LoadFromFile(std::string path, size_t num, std::string suffix) {
     img_list_.clear();
     img_list_.resize(num);
-    for (int i = 0; i < num; i++) {
+    for (size_t i = 0; i < num; i++) {
         img_list_[i] = std::make_unique<IMAGE>();
         std::string path_i = path + "_" + std::to_string(i+1) + "." + suffix;
         loadimage_safe(img_list_[i].get(), path_i.c_str());
@@ -16,7 +16,7 @@ void Atlas::LoadFromFile(std::string path, int num, std::string suffix) {
 void Atlas::LoadfromAtlas(Atlas& atlas, bool flip) {
     img_list_.clear();
     img_list_.resize(atlas.size());
-    for (int i = 0; i < atlas.size(); i++) {
+    for (size_t i = 0; i < atlas.size(); i++) {
         img_list_[i] = std::make_unique<IMAGE>();
         if (flip) flipimage(atlas.GetImage(i), img_list_[i].get());
         else *(img_list_[i]) = *(atlas.GetImage(i));
